requirements.txt
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
defusedxml==0.7.1
idna==3.3
importlib-metadata==4.8.2
jeepney==0.7.1
jira==3.1.1
keyring==23.2.1
oauthlib==3.1.1
pycodestyle==2.8.0
pycparser==2.21
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
SecretStorage==3.3.1
toml==0.10.2
urllib3==1.26.7
zipp==3.6.0
